improv: simplify BDK's connection in BDKService
