+++ BDKSwiftExampleWallet/Extensions/BDK+Extensions/Connection+Extensions.swift
@@ -1,0 +1,20 @@
+import BitcoinDevKit
+import Foundation
+
+extension Connection {
+    static func createConnection() throws -> Connection {
+        let documentsDirectoryURL = URL.documentsDirectory
+        let walletDataDirectoryURL = documentsDirectoryURL.appendingPathComponent("wallet_data")
+
+        if FileManager.default.fileExists(atPath: walletDataDirectoryURL.path) {
+            try FileManager.default.removeItem(at: walletDataDirectoryURL)
+        }
+        
+        try FileManager.default.ensureDirectoryExists(at: walletDataDirectoryURL)
+        try FileManager.default.removeOldFlatFileIfNeeded(at: documentsDirectoryURL)
+        let persistenceBackendPath = walletDataDirectoryURL.appendingPathComponent("wallet.sqlite")
+            .path
+        let connection = try Connection(path: persistenceBackendPath)
+        return connection
+    }
+}
