+++ BDKSwiftExampleWallet/Service/BDK Service/BDKService.swift
@@ -94,15 +94,12 @@ private class BDKService {
     }
     
     func createWallet(words: String?) throws {
-        let savedURL = try? keyClient.getEsploraURL()
-
-        let documentsDirectoryURL = URL.documentsDirectory
-        let walletDataDirectoryURL = documentsDirectoryURL.appendingPathComponent("wallet_data")
-
-        if FileManager.default.fileExists(atPath: walletDataDirectoryURL.path) {
-            try FileManager.default.removeItem(at: walletDataDirectoryURL)
+        self.connection = try Connection.createConnection()
+        guard let connection = connection else {
+            throw WalletError.dbNotFound
         }
-
+        
+        let savedURL = try? keyClient.getEsploraURL()
         let baseUrl = savedURL ?? network.url
 
         var words12: String
@@ -141,13 +138,7 @@ private class BDKService {
         try keyClient.saveEsploraURL(baseUrl)
         self.esploraURL = baseUrl
         updateEsploraClient()
-
-        try FileManager.default.ensureDirectoryExists(at: walletDataDirectoryURL)
-        try FileManager.default.removeOldFlatFileIfNeeded(at: documentsDirectoryURL)
-        let persistenceBackendPath = walletDataDirectoryURL.appendingPathComponent("wallet.sqlite")
-            .path
-        let connection = try Connection(path: persistenceBackendPath)
-        self.connection = connection
+        
         let wallet = try Wallet(
             descriptor: descriptor,
             changeDescriptor: changeDescriptor,
@@ -158,14 +149,19 @@ private class BDKService {
     }
 
     func createWallet(descriptor: String?) throws {
+        self.connection = try Connection.createConnection()
+        guard let connection = connection else {
+            throw WalletError.dbNotFound
+        }
+        
         let savedURL = try? keyClient.getEsploraURL()
 
-        let documentsDirectoryURL = URL.documentsDirectory
-        let walletDataDirectoryURL = documentsDirectoryURL.appendingPathComponent("wallet_data")
-
-        if FileManager.default.fileExists(atPath: walletDataDirectoryURL.path) {
-            try FileManager.default.removeItem(at: walletDataDirectoryURL)
-        }
+//        let documentsDirectoryURL = URL.documentsDirectory
+//        let walletDataDirectoryURL = documentsDirectoryURL.appendingPathComponent("wallet_data")
+//
+//        if FileManager.default.fileExists(atPath: walletDataDirectoryURL.path) {
+//            try FileManager.default.removeItem(at: walletDataDirectoryURL)
+//        }
 
         let baseUrl = savedURL ?? network.url
 
@@ -206,12 +202,12 @@ private class BDKService {
         try keyClient.saveNetwork(self.network.description)
         try keyClient.saveEsploraURL(baseUrl)
 
-        try FileManager.default.ensureDirectoryExists(at: walletDataDirectoryURL)
-        try FileManager.default.removeOldFlatFileIfNeeded(at: documentsDirectoryURL)
-        let persistenceBackendPath = walletDataDirectoryURL.appendingPathComponent("wallet.sqlite")
-            .path
-        let connection = try Connection(path: persistenceBackendPath)
-        self.connection = connection
+//        try FileManager.default.ensureDirectoryExists(at: walletDataDirectoryURL)
+//        try FileManager.default.removeOldFlatFileIfNeeded(at: documentsDirectoryURL)
+//        let persistenceBackendPath = walletDataDirectoryURL.appendingPathComponent("wallet.sqlite")
+//            .path
+//        let connection = try Connection(path: persistenceBackendPath)
+//        self.connection = connection
         let wallet = try Wallet(
             descriptor: descriptor,
             changeDescriptor: changeDescriptor,
@@ -222,14 +218,19 @@ private class BDKService {
     }
 
     func createWallet(xpub: String?) throws {
+        self.connection = try Connection.createConnection()
+        guard let connection = connection else {
+            throw WalletError.dbNotFound
+        }
+        
         let savedURL = try? keyClient.getEsploraURL()
 
-        let documentsDirectoryURL = URL.documentsDirectory
-        let walletDataDirectoryURL = documentsDirectoryURL.appendingPathComponent("wallet_data")
-
-        if FileManager.default.fileExists(atPath: walletDataDirectoryURL.path) {
-            try FileManager.default.removeItem(at: walletDataDirectoryURL)
-        }
+//        let documentsDirectoryURL = URL.documentsDirectory
+//        let walletDataDirectoryURL = documentsDirectoryURL.appendingPathComponent("wallet_data")
+//
+//        if FileManager.default.fileExists(atPath: walletDataDirectoryURL.path) {
+//            try FileManager.default.removeItem(at: walletDataDirectoryURL)
+//        }
 
         let baseUrl = savedURL ?? network.url
 
@@ -264,12 +265,12 @@ private class BDKService {
         self.esploraURL = baseUrl
         updateEsploraClient()
 
-        try FileManager.default.ensureDirectoryExists(at: walletDataDirectoryURL)
-        try FileManager.default.removeOldFlatFileIfNeeded(at: documentsDirectoryURL)
-        let persistenceBackendPath = walletDataDirectoryURL.appendingPathComponent("wallet.sqlite")
-            .path
-        let connection = try Connection(path: persistenceBackendPath)
-        self.connection = connection
+//        try FileManager.default.ensureDirectoryExists(at: walletDataDirectoryURL)
+//        try FileManager.default.removeOldFlatFileIfNeeded(at: documentsDirectoryURL)
+//        let persistenceBackendPath = walletDataDirectoryURL.appendingPathComponent("wallet.sqlite")
+//            .path
+//        let connection = try Connection(path: persistenceBackendPath)
+//        self.connection = connection
         let wallet = try Wallet(
             descriptor: descriptor,
             changeDescriptor: changeDescriptor,
@@ -438,7 +439,6 @@ private class BDKService {
         let values = wallet.sentAndReceived(tx: tx)
         return values
     }
-
 }
 
 extension BDKService {
